+++ playground/pages/index.vue
@@ -2,7 +2,6 @@
 
 <template>
   <div class="flex flex-col gap-5">
-    {{ tailwindColor("green", 500) }}
     <GButton @click="modal()">drawer</GButton>
 
     <GModal title="=moz"></GModal>