+++ src/runtime/components/G/Modal.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import { onMounted, useId, watch } from "#imports";
+import { watch } from "#imports";
 const props = defineProps({
   modelValue: {
     required: true,