+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - checkbox component added
 - radio button component added
 - avatar component added
+- alert component added
+- simple table component added
+- breadcrumb component changed props
+- drawer component now work with v-model and $drawer() not supported
 
 ## `v1.2.0`